// src/shared/types/Algorithm.ts
export type StepType =
    | 'POINTER'
    | 'COMPARE'
    | 'FOUND'
    | 'RETURN'
    | 'READ'
    | 'COMPUTE'
    | 'CREATE_NODE'
    | 'LINK'
    | 'MOVE_POINTER'
    | 'MUTATE_GRID'
    | 'DFS_CALL'
    | 'ISLAND_DISCOVERED'
    | 'SWAP'
    | 'CHECK_SORTED'
    | 'DISCARD_HALF'
    | 'UPDATE'
    | 'RESET'
    | 'INIT_SET'
    | 'ADD_TO_SET'
    | 'CHECK_START'
    | 'START_SEQUENCE'
    | 'EXPAND'
    | 'SEQUENCE_END'
    | 'UPDATE_MAX'
    | 'MARK_DONE'
    | 'SKIP'
    | 'INITIALIZE'
    | 'INIT'
    | 'SELECT'
    | 'CANDIDATES'
    | 'PROCESS'
    | 'LOOKUP'
    | 'UPDATE_MAP'
    | 'START'
    | 'COUNT_CHAR'
    | 'CHECK_FREQ'
    | 'NO_UNIQUE'
    | 'SKIP_LEFT'
    | 'SKIP_RIGHT'
    | 'MATCH'
    | 'MISMATCH'
    | 'SUCCESS'
    | 'MOVE'
    | 'START_GROUP'
    | 'COUNT'
    | 'WRITE_CHAR'
    | 'WRITE_COUNT'
    | 'PUSH'
    | 'POP'
    | 'START_WINDOW'
    | 'MOVE_WINDOW'
    | 'FAIL'
    | 'LPS_COMPUTE'
    | 'READ_STRING'
    | 'GENERATE_KEY'
    | 'MAP_LOOKUP'
    | 'CREATE_GROUP'
    | 'ADD_TO_GROUP'
    | 'SCAN'
    | 'START_WORD'
    | 'END_WORD'
    | 'PUSH_WORD'
    | 'SKIP_SPACE'
    | 'REVERSE_ORDER'
    | 'TAKE_TOP'
    | 'TAKE_LEFT'
    | 'UPDATE_CELL'
    | 'COMPARE_S1'
    | 'COMPARE_S2'
    | 'MATCH_S1'
    | 'MATCH_S2'
    | 'SET_TRUE'
    | 'SET_FALSE'
    | 'TAKE_DIAGONAL'
    | 'VISIT_NODE'
    | 'FOUND_CHILD'
    | 'DESCEND_CHILD'
    | 'REWIRE_NEXT'
    | 'REWIRE_PREV'
    | 'NULLIFY_CHILD'
    | 'RETURN_PARENT'
    | 'MOVE_FORWARD'
    | 'COMPLETE'
    | 'SET_NEXT'
    | 'REVERSE_LINK'
    | 'MOVE_PREV'
    | 'MOVE_CURR'
    | 'UPDATE_HEAD'
    | 'CALL_RECURSION'
    | 'BASE_CASE'
    | 'RETURN_FROM_CALL'
    | 'BREAK_LINK'
    | 'RETURN_HEAD'
    | 'SELECT_PIVOT'
    | 'PARTITION_DONE'
    | 'MOVE_LEFT'
    | 'MOVE_RIGHT'
    | 'NARROW_RANGE'
    | 'CHECK_CONDITION'
    | 'MOVE_SLOW'
    | 'MOVE_FAST'
    | 'FOUND_MIDDLE'
    | 'ATTACH_NODE'
    | 'MOVE_L1'
    | 'MOVE_L2'
    | 'MOVE_MERGED'
    | 'ATTACH_REMAINING'
    | 'ADVANCE_FAST'
    | 'CHECK_GAP'
    | 'MOVE_BOTH'
    | 'IDENTIFY_TARGET'
    | 'REMOVE_NODE'

export interface Step {
    type: StepType
    indices?: number[]
    rowHighlights?: Record<number, number[]>
    value?: any
    description?: string
    customValues?: Record<string, any>
    phase?: string
}